pyproject.toml
[project]
name = "daily-ai-agent"
version = "0.1.0"
description = "AI agent for morning routine briefings using MCP server tools"
readme = "README.md"
authors = [
    { name = "Kevin Reber", email = "kevinreber1@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    # Core AI framework
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    
    # HTTP client for MCP server
    "httpx>=0.25.0",
    
    # Data validation and settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.4.0",
    
    # CLI and output formatting
    "rich>=13.0.0",
    "typer>=0.9.0",
    
    # Web API framework (same as MCP server)
    "flask[async]==3.0.0",
    "flask-cors==4.0.0",
    "flask-limiter==3.5.0",
    
    # Date/time handling
    "python-dateutil>=2.8.0",
    
    # Environment variables
    "python-dotenv>=1.0.0",
    
    # Logging
    "loguru>=0.7.0",
]

[project.scripts]
daily-ai-agent = "daily_ai_agent:main"
daily-ai-agent-api = "daily_ai_agent.api_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
